+++ src/WriterTest.php
@@ -274,4 +274,34 @@ final class WriterTest extends TestCase
             ],
         ];
     }
+
+    public function testEncloseAll(): void
+    {
+        /**
+         * @see https://en.wikipedia.org/wiki/Comma-separated_values#Example
+         */
+        $records = [
+            ['Year', 'Make', 'Model', 'Description', 'Price'],
+            [1997, 'Ford', 'E350', 'ac,abs,moon', '3000.00'],
+            [1999, 'Chevy', 'Venture "Extended Edition"', null, '4900.00'],
+            [1999, 'Chevy', 'Venture "Extended Edition, Very Large"', null, '5000.00'],
+            [1996, 'Jeep', 'Grand Cherokee', 'MUST SELL!
+            air, moon roof, loaded', '4799.00'],
+        ];
+
+        $csv = Writer::createFromString();
+        $csv->setDelimiter('|');
+        $csv->forceEnclosure();
+        $csv->insertAll($records);
+
+        $expected = <<<CSV
+"Year"|"Make"|"Model"|"Description"|"Price"
+"1997"|"Ford"|"E350"|"ac,abs,moon"|"3000.00"
+"1999"|"Chevy"|"Venture ""Extended Edition"""|""|"4900.00"
+"1999"|"Chevy"|"Venture ""Extended Edition, Very Large"""|""|"5000.00"
+"1996"|"Jeep"|"Grand Cherokee"|"MUST SELL!
+            air, moon roof, loaded"|"4799.00"
+CSV;
+        self::assertStringContainsString($expected, $csv->toString());
+    }
 }